requirements.txt
PyGithub>=2.3.0
pandas>=2.1.0
pytest>=8.0.0
